chore: change received type to float

+++ app/db/models/orders.py
@@ -17,12 +17,11 @@ class Orders(Base):
     __tablename__ = "orders"
 
     id: Mapped[int] = mapped_column(primary_key=True)
-
     inv: Mapped[int] = mapped_column(nullable=False, unique=True)
     unique_code: Mapped[str] = mapped_column( nullable=False, unique=True)
     lot_type: Mapped[str] = mapped_column(nullable=False)
     buyer_email: Mapped[str] = mapped_column(nullable=False)
-    received: Mapped[int] = mapped_column(nullable=False)
+    received: Mapped[float] = mapped_column(nullable=False)
     received_currency: Mapped[str] = mapped_column(nullable=False)
     pay_time: Mapped[int] = mapped_column(nullable=False)
     check_time: Mapped[int] = mapped_column(nullable=True)
